Late updates to GitHub
I've been putting off a commit because I've been fiddling with this program a lot, now that Iconic RP sucks and I'm just kind of waiting for CircuitRP- this program doesn't need much development anymore.

Optimized pretty much all of MainWindow

This version has correct caching for items, this version also has a login system to sort users. Of course, this is on GitHub, anyone could remove the system.